outsource the drs run assignment
